+++ Assets/Mirror/CHANGELOG.md
@@ -1,3 +1,27 @@
+# [19.1.0](https://github.com/MirrorNG/MirrorNG/compare/19.0.1-master...19.1.0-master) (2020-03-19)
+
+
+### Features
+
+* Now you can pass NetworkIdentity and GameObjects ([#83](https://github.com/MirrorNG/MirrorNG/issues/83)) ([dca2d40](https://github.com/MirrorNG/MirrorNG/commit/dca2d4056fe613793480b378d25509284a1fd46a))
+
+## [19.0.1](https://github.com/MirrorNG/MirrorNG/compare/19.0.0-master...19.0.1-master) (2020-03-17)
+
+
+### Bug Fixes
+
+* calling syncvar hook when not connected yet ([#77](https://github.com/MirrorNG/MirrorNG/issues/77)) ([e64727b](https://github.com/MirrorNG/MirrorNG/commit/e64727b74bcbb1adfcd8f5efbf96066443254dff))
+
+# [19.0.0](https://github.com/MirrorNG/MirrorNG/compare/18.0.0-master...19.0.0-master) (2020-03-17)
+
+
+* removed obsoletes (#1542) ([4faec29](https://github.com/MirrorNG/MirrorNG/commit/4faec295593b81a709a57aaf374bb5b080a04538)), closes [#1542](https://github.com/MirrorNG/MirrorNG/issues/1542)
+
+
+### BREAKING CHANGES
+
+* removed obsoletes
+
 # [18.0.0](https://github.com/MirrorNG/MirrorNG/compare/17.0.2-master...18.0.0-master) (2020-03-17)